Move schema monkey patch to initializer
